Converted to the same class structure as ONNX
