+++ moonshine-tflite/README.md
@@ -1,0 +1,3 @@
+# useful-moonshine-tflite
+
+Moonshine is a family of speech-to-text models optimized for fast and accurate automatic speech recognition (ASR) on resource-constrained devices. This package contains inference code for using Moonshine models with the TensorFlow Lite runtime. For more information, please refer to the [project repo on GitHub](https://github.com/usefulsensors/moonshine).
